+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 0.3-SNAPSHOT (Unreleased)
 
+### Dependencies
+#### miop-collekt
+* **[BREAKING]** Depend on `collekt-api-*` instead of `collekt-core-*`
+* Upgrade dependency of collekt from `0.0.2` to `0.0.3`
+
 ## 0.2.0 (2018-09-23)
 Migrate to Kotlin 1.3 RC and coroutines 0.26 (Structured concurrency)