+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change log
 
-## 0.3-SNAPSHOT (Unreleased)
+## 0.3.0 (2018-12-10)
 
 ### Added
 * `produceAtomic` coroutine builder. An alternative to `produce` which guarantee that the block is always invoked.
@@ -16,6 +16,10 @@
 * **[BREAKING]** Hide dependencies on `collekt-core-*` and publish the dependency on `collekt-api-*` only (in modules `miop-java` and `miop-collekt-*`)
 * Update dependency of collekt from `0.0.2` to `0.0.3`
 
+### Experimental status
+* All channel operators have been marked with `ObsoleteCoroutinesApi`
+* Factory functions have been marked with `ExperimentalCoroutinesApi`
+   
 ## 0.2.0 (2018-09-23)
 Migrate to Kotlin 1.3 RC and coroutines 0.26 (Structured concurrency)